feat: remove unused code
