+++ p2p_wallet/Scenes/CreateOrRestoreWallet/CreateWallet/Steps/SocialSignIn/SocialSignInTryAgain/SocialSignInTryAgainViewController.swift
@@ -1,7 +1,3 @@
-// Copyright 2022 P2P Validator Authors. All rights reserved.
-// Use of this source code is governed by a MIT-style license that can be
-// found in the LICENSE file.
-
 import AuthenticationServices
 import BEPureLayout
 import Combine
