+++ p2p_wallet/Scenes/Main/ConnectionError/ConnectionErrorVC.swift
@@ -64,7 +64,7 @@ class ConnectionErrorView: BEView {
         stackView.autoPinEdgesToSuperviewEdges(with: .init(x: 20, y: 30))
 
         // separator
-        let separator = UIView.defaultSeparator()
+        let separator = UIView.separator(height: 1, color: .separator)
         view.addSubview(separator)
         separator.autoPinEdge(toSuperviewEdge: .leading)
         separator.autoPinEdge(toSuperviewEdge: .trailing)
