+++ p2p_wallet/Scenes/Main/NewSettings/NewUsernameViewController.swift
@@ -140,7 +140,7 @@ private class QrCodeCard: BECompositionView {
             }
             
             // Divider
-            UIView.defaultSeparator()
+            UIView.separator(height: 1, color: .separator)
             
             // Action buttons
             UIStackView(axis: .horizontal, alignment: .fill, distribution: .fillEqually) {
@@ -346,5 +346,20 @@ private final class PhotoLibraryAlertPresenter {
             
             UIApplication.shared.open(url, options: [:], completionHandler: nil)
         }
+    }
+}
+
+private extension UIView {
+    /// Grey banner
+    static func greyBannerView(
+        contentInset: UIEdgeInsets = .init(all: 18),
+        axis: NSLayoutConstraint.Axis = .vertical,
+        spacing: CGFloat = 8,
+        alignment: UIStackView.Alignment = .fill,
+        distribution: UIStackView.Distribution = .fill,
+        @BEStackViewBuilder builder: () -> [BEStackViewElement]
+    ) -> UIView {
+        UIStackView(axis: axis, spacing: spacing, alignment: alignment, distribution: distribution, builder: builder)
+            .padding(contentInset, backgroundColor: .a3a5ba.withAlphaComponent(0.05), cornerRadius: 12)
     }
 }
